--- src/cudapatch.jl
@@ -1,99 +1,0 @@
-using .CUDA
-
-using TupleTools
-using Base.Cartesian
-import .CUDA: @cuindex
-
-# patch file for CUDA.
-function CUDA.mapreducedim_kernel_parallel(f, op, R::CuDeviceArray{T}, A,
-                             CIS, Rlength, Slength) where {T}
-    for Ri_base in 0:(gridDim().x * blockDim().y):(Rlength-1)
-        Ri = Ri_base + (blockIdx().x - 1) * blockDim().y + threadIdx().y
-        Ri > Rlength && return
-        RI = Tuple(CartesianIndices(R)[Ri])
-        S = @cuStaticSharedMem(T, 512)
-        Si_folded_base = (threadIdx().y - 1) * blockDim().x
-        Si_folded = Si_folded_base + threadIdx().x
-        # serial reduction of A into S by Slength ÷ xthreads
-        for Si_base in 0:blockDim().x:(Slength-1)
-            Si = Si_base + threadIdx().x
-            Si > Slength && break
-            SI = Tuple(CIS[Si])
-            AI = ifelse.(size(R) .== 1, SI, RI)
-            if Si_base == 0
-                S[Si_folded] = f(A[AI...])
-            else
-                S[Si_folded] = op(S[Si_folded], f(A[AI...]))
-            end
-        end
-        # block-parallel reduction of S to S[1] by xthreads
-        CUDA.reduce_block(view(S, (Si_folded_base + 1):512), op)
-        # reduce S[1] into R
-        threadIdx().x == 1 && (R[Ri] = op(R[Ri], S[Si_folded]))
-    end
-    return
-end
-
-function Base._mapreducedim!(f, op, R::CuArray{T}, A::EinArray{T}) where {T}
-    # the kernel as generated from `f` and `op` can require lots of registers (eg. #160),
-    # so we need to be careful about how many threads we launch not to run out of them.
-    Rlength = length(R)
-    Ssize = ifelse.(size(R) .== 1, size(A), 1)
-    Slength = prod(Ssize)
-    CIS = CartesianIndices(Ssize)
-
-    parallel_args = (f, op, R, A, CIS, Rlength, Slength)
-    # NOTE: why is GC.@preserve ?
-    GC.@preserve parallel_args begin
-        # NOTE: why not using `@cuda` here?
-        parallel_kargs = cudaconvert.(parallel_args)  # CuArray -> DevicePtr
-        parallel_tt = Tuple{Core.Typeof.(parallel_kargs)...}
-        parallel_kernel = cufunction(CUDA.mapreducedim_kernel_parallel, parallel_tt)
-
-        # we are limited in how many threads we can launch...
-        ## by the kernel
-        kernel_threads = CUDA.maxthreads(parallel_kernel)
-        ## by the device
-        dev = CUDAdrv.device()
-        block_threads = (x=attribute(dev, CUDAdrv.DEVICE_ATTRIBUTE_MAX_BLOCK_DIM_X),
-                         y=attribute(dev, CUDAdrv.DEVICE_ATTRIBUTE_MAX_BLOCK_DIM_Y),
-                         total=attribute(dev, CUDAdrv.DEVICE_ATTRIBUTE_MAX_THREADS_PER_BLOCK))
-
-        # figure out a legal launch configuration
-        y_thr = min(nextpow(2, Rlength ÷ 512 + 1), 512, block_threads.y, kernel_threads)
-        x_thr = min(512 ÷ y_thr, Slength, block_threads.x,
-                    ceil(Int, block_threads.total/y_thr),
-                    ceil(Int, kernel_threads/y_thr))
-
-        if x_thr >= 8
-            blk, thr = (Rlength - 1) ÷ y_thr + 1, (x_thr, y_thr, 1)
-            parallel_kernel(parallel_kargs...; threads=thr, blocks=blk)
-        else
-            # not enough work, fall back to serial reduction
-            range = ifelse.(length.(axes(R)) .== 1, axes(A), nothing)
-            blk, thr = CUDA.cudims(R)
-            @cuda(blocks=blk, threads=thr, CUDA.mapreducedim_kernel_serial(f, op, R, A, range))
-        end
-    end
-
-    return R
-end
-
-using .CUDA: @cartesianidx, AbstractGPUArray, gpu_call
-
-Base.ndims(::Base.Broadcast.Broadcasted{CUDA.CuArrayStyle{0}}) = 0
-
-@inline @generated function map_index(dest, src, I, perm::NTuple{N,T}) where {N,T}
-    Expr(:(=), Expr(:ref, :dest, [:(@inbounds I[perm[$i]]) for i in 1:N]...), Expr(:ref, :src, :I))
-end
-function LinearAlgebra.permutedims!(dest::AbstractGPUArray, src::AbstractGPUArray,
-                                    perm::NTuple)
-    Base.checkdims_perm(dest, src, perm)
-    function permutedims_kernel(ctx, dest, src, perm)
-        I = @cartesianidx src
-        map_index(dest, src, I, perm)
-        return
-    end
-    gpu_call(permutedims_kernel, dest, src, perm)
-    return dest
-end
