+++ src/cueinsum.jl
@@ -81,4 +81,23 @@ function einsum(::SimpleBinaryRule{(),(), ()}, xs::NTuple{2, DenseCuArray})
     asarray(Array(xs[1])[] * Array(xs[2])[], xs[1])
 end
 
+using .CUDA: @cartesianidx, AbstractGPUArray, gpu_call
+
+Base.ndims(::Base.Broadcast.Broadcasted{CUDA.CuArrayStyle{0}}) = 0
+
+@inline @generated function map_index(dest, src, I, perm::NTuple{N,T}) where {N,T}
+    Expr(:(=), Expr(:ref, :dest, [:(@inbounds I[perm[$i]]) for i in 1:N]...), Expr(:ref, :src, :I))
+end
+function LinearAlgebra.permutedims!(dest::AbstractGPUArray, src::AbstractGPUArray,
+                                    perm::NTuple)
+    Base.checkdims_perm(dest, src, perm)
+    function permutedims_kernel(ctx, dest, src, perm)
+        I = @cartesianidx src
+        map_index(dest, src, I, perm)
+        return
+    end
+    gpu_call(permutedims_kernel, dest, src, perm)
+    return dest
+end
+
 @info("OMEinsum loaded the CUDA module successfully")
